% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlmodels-utils.R
\name{BIC}
\alias{BIC}
\title{Extract BIC}
\usage{
BIC(object, ...)
}
\arguments{
\item{object}{An object of class \code{"mlmodel"} or \code{"summary.mlmodel"}.}

\item{...}{Further arguments passed to methods.}
}
\description{
Extract BIC
}
