% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlmodel-utils.R
\name{coef.mlmodel}
\alias{coef.mlmodel}
\title{Extract Model Coefficients}
\usage{
\method{coef}{mlmodel}(object, ...)
}
\arguments{
\item{object}{An \code{mlmodel} object.}

\item{...}{Currently not used.}
}
\value{
A named numeric vector of estimated coefficients.
}
\description{
Extract Model Coefficients
}
