% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlmodel-utils.R
\name{confint.mlmodel}
\alias{confint.mlmodel}
\title{Gets the confidence interval for an estimated model.}
\usage{
\method{confint}{mlmodel}(
  object,
  parm,
  level = 0.95,
  vcov = NULL,
  vcov.type = "oim",
  cl_var = NULL,
  repetitions = 999,
  seed = NULL,
  progress = FALSE,
  ...
)
}
\arguments{
\item{object}{Any model estimated with one of our estimators.}

\item{parm}{A vector with the names or indices of the parameters you want the
interval for. If missing all parameters are considered.}

\item{level}{The confidence level for the interval.}

\item{vcov}{Optional user-supplied variance-covariance matrix.}

\item{vcov.type}{Type of variance-covariance matrix. See \link[=vcov.mlmodel]{vcov}.}

\item{cl_var}{Clustering variable (name or vector).}

\item{repetitions}{Number of bootstrap replications when \code{vcov.type = "boot"}.}

\item{seed}{Random seed for the boostrapping, for reproducibility.}

\item{progress}{Logical. Show bootstrap/jackknife progress bar? Default is
\code{FALSE} in higher-level functions.}

\item{...}{Additional arguments for methods.}
}
\description{
Gets the confidence interval for an estimated model.
}
