% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detect_log_transformation}
\alias{detect_log_transformation}
\title{Detect log transformation on the outcome variable}
\usage{
detect_log_transformation(formula, data)
}
\description{
Returns a consistent list structure whether a log transformation is detected or not.
}
\keyword{internal}
