% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{docvis}
\alias{docvis}
\title{U.S. Medical Expenditure Panel Survey}
\format{
A data frame with 3677 rows and 22 variables.
\describe{
\item{offer}{Binary: employer offers insurance}
\item{ssiratio}{Ratio of SSI income to total income}
\item{age}{Age in years}
\item{educyr}{Years of education}
\item{physician}{Number of visits to doctor}
\item{nonphysician}{Number of visits to health professional (not doctor)}
\item{medicaid}{Binary: has medicaid public insurance}
\item{private}{Binary: has private supplementary insurance}
\item{female}{Binary: female}
\item{phylim}{Binary: physical limitation}
\item{actlim}{Binary: activity limitation}
\item{income}{Income (in thousands)}
\item{totchr}{Number of chronic conditions}
\item{insured}{Same as \code{private}}
\item{age2}{\code{age} squared}
\item{linc}{\code{log(income)}}
\item{bh}{Binary: black or hispanic}
\item{docvis}{Number of doctor visits}
\item{ldocvis}{\code{log(docvis)} if \code{docvis > 0}}
\item{ldocvisa}{\code{log(docvis + .01)}}
\item{one}{Constant term (1)}
\item{docbin}{Binary: \code{docvis > 0}}
}
}
\source{
\url{https://www.stata-press.com/data/mus2.html} \code{mus220mepsdocvis.dta}
}
\usage{
docvis
}
\description{
Cross sectional sample for 2003.
}
\references{
Cameron, A. C., and P. K. Trivedi. 2022. \emph{Microeconometrics Using Stata,
Second Edition}. Volumes I and II. College Station, TX: Stata Press.
}
\keyword{datasets}
