% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlmodels-auxiliary.R
\name{.convert_integers_to_double}
\alias{.convert_integers_to_double}
\title{Function to convert variables stored as integers in a data frame into
doubles. To be called before using hardhat::mold in the estimator's function.}
\usage{
.convert_integers_to_double(data)
}
\description{
Function to convert variables stored as integers in a data frame into
doubles. To be called before using hardhat::mold in the estimator's function.
}
\keyword{internal}
