% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlmodel-auxiliary.R
\name{.is_invertible}
\alias{.is_invertible}
\title{Internal function to detect if a matrix is invertible.}
\usage{
.is_invertible(matrix)
}
\arguments{
\item{matrix}{The matrix you want to check}
}
\value{
Logical with \code{TRUE} if it's invertible, or \code{FALSE} if it's not.
}
\description{
Internal function to detect if a matrix is invertible.
}
\details{
Used in prediction and other methods to avoid throwing an error.
}
\keyword{internal}
