% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma-fit.R
\name{.ml_gamma.fit}
\alias{.ml_gamma.fit}
\title{Internal fitting function for ml_lm}
\usage{
.ml_gamma.fit(
  y,
  x,
  z,
  w = NULL,
  method = "NR",
  start = NULL,
  constraints = NULL,
  control = list(tol = -1, reltol = 1e-12, gradtol = 1e-12, lambdatol = 1e-20, qac =
    "marquardt"),
  ...
)
}
\arguments{
\item{y}{Numeric vector of outcomes.}

\item{x}{Design matrix for the value equation.}

\item{z}{Design matrix for the scale equation (or single column of 1s).}

\item{w}{Numeric vector of weights.}

\item{start}{Numeric vector of initial values.}

\item{constraints}{Parsed constraints (internal use).}

\item{control}{Control list passed to \code{\link[maxLik:maxLik]{maxLik::maxLik()}}.}

\item{...}{Further arguments passed to \code{\link[maxLik:maxLik]{maxLik::maxLik()}}.}
}
\description{
Internal fitting function for ml_lm
}
\keyword{internal}
