% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm-fit.R
\name{.ml_lm.fit}
\alias{.ml_lm.fit}
\title{Internal fitting function for ml_lm}
\usage{
.ml_lm.fit(...)
}
\arguments{
\item{...}{Further arguments passed to \code{\link[maxLik:maxLik]{maxLik::maxLik()}}.}

\item{y}{Numeric vector of outcomes.}

\item{x}{Design matrix for the value equation.}

\item{z}{Design matrix for the scale equation (or single column of 1s).}

\item{w}{Numeric vector of weights.}

\item{constraints}{Parsed constraints (internal use).}

\item{control}{Control list passed to \code{\link[maxLik:maxLik]{maxLik::maxLik()}}.}
}
\description{
Internal fitting function for ml_lm
}
\keyword{internal}
