% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit-fit.R
\name{.ml_logit.fit}
\alias{.ml_logit.fit}
\title{Stripped down function to estimate a binary logit model.}
\usage{
.ml_logit.fit(
  y,
  x,
  z = NULL,
  w = NULL,
  constraints = NULL,
  start = NULL,
  method = "NR",
  control = list(tol = -1, reltol = 1e-12, gradtol = 1e-12, lambdatol = 1e-20, qac =
    "marquardt"),
  ...
)
}
\description{
Stripped down function to estimate a binary logit model.
}
\keyword{internal}
