% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negbin-fit.R
\name{.ml_negbin.fit}
\alias{.ml_negbin.fit}
\title{Internal fitting function for ml_negbin}
\usage{
.ml_negbin.fit(
  y,
  x,
  z,
  w = NULL,
  method = "NR",
  start = NULL,
  constraints = NULL,
  dispersion = "NB2",
  control = list(tol = -1, reltol = 1e-12, gradtol = 1e-12, lambdatol = 1e-20, qac =
    "marquardt"),
  ...
)
}
\description{
Internal fitting function for ml_negbin
}
\keyword{internal}
