% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negbin-auxiliary.R
\name{.ml_negbin_nb1_gradientObs}
\alias{.ml_negbin_nb1_gradientObs}
\title{Gradient by Observations for an \code{ml_negbin} model with NB1 dispersion.}
\usage{
.ml_negbin_nb1_gradientObs(object)
}
\arguments{
\item{object}{An \code{ml_negbin} object with \code{object$model$dispersion = 'NB1'}}
}
\description{
Gradient by Observations for an \code{ml_negbin} model with NB1 dispersion.
}
\details{
The difference between this function and \code{.ml_negbin_nb1_score} is that this
one pulls the data used in estimation from \code{object$model}, and that one you
feed the data directly into the function.

Usually there is no need for this function, because \link[maxLik:maxLik]{maxLik}
stores this in \code{gradientObs}, so it will be at the root of the object.
Programmed for consistency since the \code{hessianObs} and \code{loglikeObs} functions
were programmed for use with different tests.
}
\keyword{internal}
