% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negbin-auxiliary.R
\name{.ml_negbin_nb2_loglikeObs}
\alias{.ml_negbin_nb2_loglikeObs}
\title{Log-likelihood by Observations of an NB2 model.}
\usage{
.ml_negbin_nb2_loglikeObs(object)
}
\arguments{
\item{object}{An object of class \code{ml_negbin} and \code{object$model$dispersion = 'NB2'}.}
}
\value{
Vector with the log-likelihoods, with length equal to the length of
the sample used in estimation.
}
\description{
Extracts the log-likelihood by observation from a model estimated by
\code{ml_negbin} with \code{dispersion = 'NB2'}.
}
\details{
For an estimated model it extracts the log-likelihood by observations.
}
\keyword{internal}
