% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negbin-auxiliary.R
\name{.ml_negbin_nb2_score}
\alias{.ml_negbin_nb2_score}
\title{Scores of an NB2 model}
\usage{
.ml_negbin_nb2_score(b, y, x, z, w = NULL)
}
\arguments{
\item{b}{Vector of coefficients.}

\item{y}{Vector of the outcome variable.}

\item{x}{Matrix with predictors of value equation.}

\item{z}{Matrix with predictors of scale (dispersion) equation.}

\item{w}{Vector with weights for estimation.}
}
\description{
Scores of an NB2 model at for a specific set of data and vector of coefficients.
}
\details{
Programmed to use from within \code{predict.ml_negbin}, to get the gradient of the
probabilities for the delta method standard errors of those predictions.
}
\keyword{internal}
