% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson-fit.R
\name{.ml_poisson.fit}
\alias{.ml_poisson.fit}
\title{Internal fitting function for ml_poisson}
\usage{
.ml_poisson.fit(
  y,
  x,
  w = NULL,
  method = "NR",
  start = NULL,
  constraints = NULL,
  control = list(tol = -1, reltol = 1e-12, gradtol = 1e-12, lambdatol = 1e-20, qac =
    "marquardt"),
  ...
)
}
\arguments{
\item{y}{Numeric vector of outcomes.}

\item{x}{Design matrix for the value equation.}

\item{w}{Numeric vector of weights.}

\item{start}{Numeric vector of initial values.}

\item{constraints}{Parsed constraints (internal use).}

\item{control}{Control list passed to \code{\link[maxLik:maxLik]{maxLik::maxLik()}}.}

\item{...}{Further arguments passed to \code{\link[maxLik:maxLik]{maxLik::maxLik()}}.}
}
\description{
Internal fitting function for ml_poisson
}
\keyword{internal}
