% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probit-fit.R
\name{.ml_probit.fit}
\alias{.ml_probit.fit}
\title{Stripped down function to estimate a binary logit model.}
\usage{
.ml_probit.fit(
  y,
  x,
  z = NULL,
  w = NULL,
  constraints = NULL,
  start = NULL,
  method = NULL,
  control = NULL,
  ...
)
}
\description{
Stripped down function to estimate a binary logit model.
}
\keyword{internal}
