% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlmodels-auxiliary.R
\name{.mlmodels_get_default}
\alias{.mlmodels_get_default}
\title{Get default value for mlmodels}
\usage{
.mlmodels_get_default(name)
}
\arguments{
\item{name}{Character. Name of the option.}
}
\description{
Get default value for mlmodels
}
\keyword{internal}
