% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlmodels-auxiliary.R
\name{.mold_fix_integer_to_double}
\alias{.mold_fix_integer_to_double}
\title{The function takes a mold (object returned by hardhat::mold), loops through
the ptypes in its blueprint and changes the integer types in the blueprint
to double. Then returns the changed mold.}
\usage{
.mold_fix_integer_to_double(mold)
}
\description{
The function takes a mold (object returned by hardhat::mold), loops through
the ptypes in its blueprint and changes the integer types in the blueprint
to double. Then returns the changed mold.
}
\keyword{internal}
