% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlmodel-auxiliary.R
\name{.parse_string_constraints}
\alias{.parse_string_constraints}
\title{Internal helper: parse string constraints into maxLik matrices}
\usage{
.parse_string_constraints(strings, coef_names)
}
\arguments{
\item{strings}{A vector with the strings defining the linear constraints.}

\item{coef_names}{Character vector of all coefficient names in the model.}
}
\value{
A list that conforms with \link[maxLik:maxLik]{maxLik}'s constraint
requirements.
}
\description{
Internal helper: parse string constraints into maxLik matrices
}
\keyword{internal}
