% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlmodels-auxiliary.R
\name{.predict_types_parsing}
\alias{.predict_types_parsing}
\title{Internal parser for probability types.}
\usage{
.predict_types_parsing(type)
}
\arguments{
\item{type}{Character string passed by the user}
}
\value{
A list with components:
\itemize{
\item base_type: "link", "response", or "prob"
\item prob_type: NULL, "exact", "leq", "geq", or "interval"
\item lower: lower bound (NULL if not applicable)
\item upper: upper bound (NULL if not applicable)
}
}
\description{
Internal parser for probability types.
}
\details{
\code{prob_type} can only be NULL when base_type is not "prob". So in the probability
block of the prediction function, you don't need to account for NULL.

If \code{prob_type} is "exact" both lower and upper hold the value. So you can
use either.
}
\keyword{internal}
