% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit-auxiliary.R
\name{.vcov_boot.ml_logit}
\alias{.vcov_boot.ml_logit}
\title{Bootstrap variance-covariance for ml_logit objects}
\usage{
\method{.vcov_boot}{ml_logit}(
  object,
  repetitions = 999,
  seed = NULL,
  cl_var = NULL,
  progress = TRUE,
  ...
)
}
\description{
Internal function called by vcov.mlmodel when type = "boot".
Uses .ml_logit.fit directly for efficiency and control.
}
\keyword{internal}
