% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negbin-auxiliary.R
\name{.vcov_boot.ml_negbin}
\alias{.vcov_boot.ml_negbin}
\title{Bootstrap variance-covariance matrix for ml_nebin models}
\usage{
\method{.vcov_boot}{ml_negbin}(
  object,
  repetitions = 999,
  seed = NULL,
  cl_var = NULL,
  progress = TRUE,
  ...
)
}
\arguments{
\item{object}{A fitted \code{ml_nebin} object.}

\item{repetitions}{Number of bootstrap replications.}

\item{seed}{Random seed for reproducibility.}

\item{cl_var}{Clustering variable (if doing clustered bootstrap).}

\item{progress}{Logical. Whether to show a progress bar.}

\item{...}{Not currently used.}
}
\description{
Specialized (faster) bootstrap method for \code{ml_lm} objects. It calls
\code{.ml_negbin.fit()} directly instead of going through the generic \code{update()}
method.
}
\keyword{internal}
