% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson-auxiliary.R
\name{.vcov_boot.ml_poisson}
\alias{.vcov_boot.ml_poisson}
\title{Bootstrap variance-covariance for ml_poisson objects}
\usage{
\method{.vcov_boot}{ml_poisson}(
  object,
  repetitions = 999,
  seed = NULL,
  cl_var = NULL,
  progress = TRUE,
  ...
)
}
\description{
Internal function called by vcov.mlmodel when type = "boot".
Uses .ml_probit.fit directly for efficiency and control.
}
\keyword{internal}
