% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlmodels-auxiliary.R
\name{.vcov_boot}
\alias{.vcov_boot}
\title{Generic for bootstrap variance-covariance matrix}
\usage{
.vcov_boot(object, ...)
}
\description{
Generic for bootstrap variance-covariance matrix
}
\keyword{internal}
