% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlmodels-auxiliary.R
\name{.vcov_cluster_info}
\alias{.vcov_cluster_info}
\title{Internal helper to extract clustering information}
\usage{
.vcov_cluster_info(object, cl_var)
}
\arguments{
\item{object}{An \code{mlmodel} object.}

\item{cl_var}{Character string or vector. The clustering variable.}
}
\value{
A list with \code{var_name}, \code{n_cluster}, and \code{ids}.
}
\description{
Used by several functions to that accept \code{cl_var} as an argument, to check the
validity of the clustering variable, and get the number of clusters, as well
as the name of the clustering variable if \code{cl_var} is a string.
}
\keyword{internal}
