% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm-auxiliary.R, R/logit-auxiliary.R,
%   R/mlmodels-auxiliary.R
\name{.vcov_jack.ml_lm}
\alias{.vcov_jack.ml_lm}
\alias{.vcov_jack.ml_logit}
\alias{.vcov_jack}
\alias{.vcov_jack.mlmodel}
\title{Jackknife Variance-Covariance Matrix (mlmodel method)}
\usage{
\method{.vcov_jack}{ml_lm}(object, cl_var = NULL, progress = TRUE, ...)

\method{.vcov_jack}{ml_logit}(object, cl_var = NULL, progress = TRUE, ...)

.vcov_jack(object, ...)

\method{.vcov_jack}{mlmodel}(object, cl_var = NULL, progress = TRUE, ...)
}
\arguments{
\item{object}{An \code{mlmodel} object.}

\item{cl_var}{Clustering variable (if clustered jackknife).}

\item{progress}{Logical. Whether to show progress bar.}

\item{...}{Not currently used.}
}
\description{
Internal function to compute a jackknife variance-covariance matrix.
}
\details{
Called by \link[=vcov]{vcov} when \code{type} is set to \code{"jack"}
}
\keyword{internal}
