% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginaleffects-compat.R
\name{find_predictors.mlmodel}
\alias{find_predictors.mlmodel}
\title{Extract the predictors used in the model (for insight/marginaleffects compatibility)}
\usage{
\method{find_predictors}{mlmodel}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"mlmodel"}}

\item{...}{Further arguments passed to methods}
}
\value{
A character vector with the names of the predictor variables.
}
\description{
Extract the predictors used in the model (for insight/marginaleffects compatibility)
}
