% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginaleffects-compat.R
\name{find_variables.mlmodel}
\alias{find_variables.mlmodel}
\title{Extract the variables used in the model (for insight/marginaleffects compatibility)}
\usage{
find_variables.mlmodel(x, ...)
}
\arguments{
\item{x}{An object of class \code{"mlmodel"}}

\item{...}{Further arguments passed to methods}
}
\value{
A list with components \code{response} (character) and \code{conditional} (character vector).
}
\description{
Extract the variables used in the model (for insight/marginaleffects compatibility)
}
