% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm-postestimation.R
\name{fitted.ml_lm}
\alias{fitted.ml_lm}
\title{Get the fitted value from a linear model estimated with \link{ml_lm}.}
\usage{
\method{fitted}{ml_lm}(object, ...)
}
\arguments{
\item{object}{A model class \code{ml_lm}, estimated by \link{ml_lm}.}

\item{...}{Currently not used.}
}
\value{
A vector with the fitted values.
}
\description{
Get the fitted value from a linear model estimated with \link{ml_lm}.
}
\author{
Alfonso Sanchez-Penalver
}
