% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson-postestimation.R
\name{fitted.ml_poisson}
\alias{fitted.ml_poisson}
\title{Extract Fitted Values from ml_poisson objects}
\usage{
\method{fitted}{ml_poisson}(object, ...)
}
\arguments{
\item{object}{A fitted \code{ml_poisson} object.}

\item{...}{Not currently used.}
}
\value{
A numeric vector of fitted values with length equal to the number of
observations used in estimation.
}
\description{
Returns the fitted values (predicted conditional mean) from the value equation.
}
\author{
Alfonso Sanchez-Penalver
}
