% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlmodels-utils.R
\name{fitted.mlmodel}
\alias{fitted.mlmodel}
\alias{fitted.values.mlmodel}
\title{Returns the fitted values from an object of class \code{mlmodel}.}
\usage{
\method{fitted}{mlmodel}(object, ...)

\method{fitted}{values.mlmodel}(object, ...)
}
\arguments{
\item{object}{Object of class \code{mlmodel} estmated with one of the models in
package.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A vector with the fitted values of the model.
}
\description{
Returns the fitted values from an object of class \code{mlmodel}.
}
