% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlmodel-utils.R
\name{formula.mlmodel}
\alias{formula.mlmodel}
\title{Extract value formula from mlmodel objects}
\usage{
\method{formula}{mlmodel}(x, ...)
}
\arguments{
\item{x}{An \code{mlmodel} object}

\item{...}{Currently not implemented}
}
\value{
The formula of the value equation (object of class \code{formula}).
}
\description{
Extract value formula from mlmodel objects
}
