% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginaleffects-compat.R
\name{get_coef.mlmodel}
\alias{get_coef.mlmodel}
\title{Implementation of \code{marginaleffects} package's \code{get_coef()} function.}
\usage{
get_coef.mlmodel(model, ...)
}
\arguments{
\item{model}{An \code{mlmodel} estimation object (from one of the package's
estimation functions)}

\item{...}{Additional arguments sent to methods.}
}
\value{
A named vector with the coefficients in the model.
}
\description{
Implementation of \code{marginaleffects} package's \code{get_coef()} function.
}
