% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginaleffects-compat.R
\name{get_data.mlmodel}
\alias{get_data.mlmodel}
\alias{get_modeldata.mlmodel}
\title{Extract data used to fit the model (for insight/marginaleffects compatibility)}
\usage{
\method{get_data}{mlmodel}(x, ...)

get_modeldata.mlmodel(x, ...)
}
\arguments{
\item{x}{An object of class \code{"mlmodel"}}

\item{...}{Further arguments (currently ignored)}
}
\value{
The original data frame used when fitting the model
}
\description{
Extract data used to fit the model (for insight/marginaleffects compatibility)
}
