% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginaleffects-compat.R
\name{get_predict.mlmodel}
\alias{get_predict.mlmodel}
\title{Implementation of \code{marginaleffects} package's \code{get_predict()} function.}
\usage{
get_predict.mlmodel(model, newdata = NULL, type = "response", ...)
}
\arguments{
\item{model}{An \code{mlmodel} estimation object (from one of the package's
estimation functions)}

\item{newdata}{A data.frame with the values of the predictors you want to use.
It defaults to \code{NULL}, and in such case, it will use the data used in the
estimation for the predictions (in sample prediction).}

\item{type}{A string with the type of prediction you want. This will depend
on the types available for the model you estimated. See
\link[=predict.mlmodel]{predict}.}

\item{...}{Additional arguments sent to methods.}
}
\value{
A data.frame with two variables: \code{rowid} with the row number, and
\code{estimate} with the value of the prediction requested.
}
\description{
Implementation of \code{marginaleffects} package's \code{get_predict()} function.
}
