% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginaleffects-compat.R
\name{get_vcov}
\alias{get_vcov}
\alias{get_vcov.mlmodel}
\title{Implementation of \code{marginaleffects} package's \code{get_vcov()} function.}
\usage{
get_vcov(model, ...)

\method{get_vcov}{mlmodel}(model, ...)
}
\arguments{
\item{model}{An \code{mlmodel} estimation object (from one of the package's
estimation functions)}

\item{...}{Additional arguments passed to \link[=vcov.mlmodel]{vcov}}
}
\value{
A matrix from the package's \code{vcov()} function. See
\verb{[vcov][mlmodels::vcov.mlmodel]} for more details.
}
\description{
Implementation of \code{marginaleffects} package's \code{get_vcov()} function.
}
