% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlmodel-utils.R
\name{hessianObs}
\alias{hessianObs}
\alias{hessianObs.mlmodel}
\title{Hessian by Observation}
\usage{
hessianObs(object)

\method{hessianObs}{mlmodel}(object)
}
\arguments{
\item{object}{An \code{mlmodel} object.}
}
\value{
A numeric matrix of dimension \verb{(N*K) x K}, where \code{N} is the number
of observations and \code{K} is the number of parameters. The Hessian for each
observation is stacked vertically.
}
\description{
Extract the per-observation Hessian matrices evaluated at the estimated
parameters from an \code{mlmodel} object.
}
\details{
This is mainly intended for advanced use (e.g., custom diagnostics or
information matrix tests). For most users, the functions \code{IMtest} or
\code{vcov} are more convenient.
}
