% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlmodels-utils.R
\name{logLik}
\alias{logLik}
\alias{logLik.mlmodel}
\alias{logLik.summary.mlmodel}
\title{Return the Log-likelihood value.}
\usage{
logLik(object, ...)

\method{logLik}{mlmodel}(object, ...)

\method{logLik}{summary.mlmodel}(object, ...)
}
\arguments{
\item{object}{Object of class \code{mlmodel} or \code{summary.mlmodel}, estimated with
one of the estimators in this package.}

\item{...}{Additional arguments to methods.}
}
\value{
A scalar numeric: the log-likelihood of the model. It includes the
number of observations as the attribute 'nobs'.
}
\description{
Return the Log-likelihood value.
}
