% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlmodel-utils.R
\name{loglikeObs.mlmodel}
\alias{loglikeObs.mlmodel}
\title{Log-Likelihood by observation}
\usage{
\method{loglikeObs}{mlmodel}(object)
}
\arguments{
\item{object}{An object of class \code{mlmodel}, i.e. an model estimated with one
of the estimators in the package.}
}
\value{
Vector with the evaluated log-likelihoods per observation.
}
\description{
Get the log-likelihood, evaluated at the optimum from an object of class
\code{mlmodel}.
}
\author{
Alfonso Sanchez-Penalver
}
