% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm-fit.R
\name{ml_lm}
\alias{ml_lm}
\title{Fit linear model by Maximum Likelihood}
\usage{
ml_lm(
  value,
  scale = NULL,
  weights = NULL,
  data,
  subset = NULL,
  noint_value = FALSE,
  noint_scale = FALSE,
  constraints = NULL,
  control = NULL,
  ...
)
}
\arguments{
\item{value}{Formula for the conditional mean (value) equation.}

\item{scale}{Formula for log(sigma) (optional). If NULL, a homoskedastic
model is fitted.}

\item{weights}{Optional weights variable. It can be either the name of the
variable in \code{data}, or a vector with the weights.}

\item{data}{Data frame.}

\item{subset}{Optional subset expression. Only observations for which this
expression evaluates to \code{TRUE} are used in the estimation. This can be
a logical vector or an expression (e.g. \code{subset = age > 30}).}

\item{noint_value}{Logical. Should the value equation omit the intercept?
Default is \code{FALSE}.}

\item{noint_scale}{Logical. Should the scale equation omit the intercept?
Default is \code{FALSE}.}

\item{control}{A list of control parameters passed to \code{\link[maxLik:maxLik]{maxLik::maxLik()}}.
The default values are chosen to work well with difficult likelihoods.
See \code{\link[maxLik:maxLik]{maxLik::maxLik()}} for details.}

\item{...}{Additional arguments passed to \code{\link[maxLik:maxLik]{maxLik::maxLik()}}.}
}
\value{
An object of class \code{ml_lm} that extends \code{mlmodel}.
}
\description{
Fit linear model by Maximum Likelihood
}
\details{
\strong{Important:} Do not use the usual R syntax to remove the intercept in the
formulas (\code{- 1} or \code{+ 0}). Use the dedicated arguments \code{noint_value} and
\code{noint_scale} instead.
}
\examples{
if (requireNamespace("wooldridge", quietly = TRUE)) {

  # Homoskedastic model
  linm <- ml_lm(cigs ~ log(income) + age + I(age^2) + educ + restaurn,
                data = wooldridge::smoke)
  summary(linm)

  # Heteroskedastic model
  linm_het <- ml_lm(value = cigs ~ log(income) + age + I(age^2) + educ + restaurn,
                    scale = ~ age + I(age^2) + restaurn,
                    data = wooldridge::smoke)
  summary(linm_het, vcov.type = "robust")
}

}
\author{
Alfonso Sanchez-Penalver
}
