% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlmodels-package.R
\docType{package}
\name{mlmodels-package}
\alias{mlmodels}
\alias{mlmodels-package}
\title{mlmodels: Maximum Likelihood Models and Tools for Estimation, Prediction, and Testing}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a collection of maximum likelihood estimators with a consistent S3 interface. Supported models include Gaussian (linear and log-normal), logit, probit, Poisson, negative binomial (NB1 and NB2), gamma, and beta regression. A distinctive feature is flexible modeling of the scale parameter (variance, dispersion, precision, or shape) alongside the location/mean parameters. The package offers unified predict() methods, multiple variance-covariance estimators (observed information, outer product of gradients, robust/Huber-White, cluster-robust, bootstrap, jackknife), and a full suite of hypothesis tests (Wald, likelihood ratio, information matrix, Vuong, overdispersion, and goodness-of-fit). It is fully compatible with 'marginaleffects' for post-estimation analysis. Methods implemented include Cameron and Trivedi (1990) \doi{10.1016/0304-4076(90)90014-K}, for Poisson overdispersion testing, Manjon and Martinez (2014) \doi{10.1177/1536867X1401400406}, for goodness-of-fit testing of count data models, Vuong (1989) \doi{10.2307/1912557}, for non-nested likelihood ratio testing, and White (1982) \doi{10.2307/1912526}, for information matrix tests.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://alfisankipan.github.io/mlmodels/}
  \item Report bugs at \url{https://github.com/alfisankipan/mlmodels/issues}
}

}
\author{
\strong{Maintainer}: Alfonso Sanchez-Penalver \email{oneiros_spain@yahoo.com} (\href{https://orcid.org/0000-0001-8491-4632}{ORCID})

}
\keyword{internal}
