% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mroz}
\alias{mroz}
\title{University of Michigan Panel Study of Income Dynamics (PSID)}
\format{
A data frame with 753 observations on 22 variables.
\describe{
\item{inlf}{Binary: in labor force in 1975}
\item{hours}{Hours worked in 1975}
\item{kidslt6}{Number of children less than 6 years old}
\item{kidsge6}{Number of children 6 years old or older}
\item{age}{Woman's age in years}
\item{educ}{Woman's years of education}
\item{wage}{Woman's estimated average hourly earnings}
\item{repwage}{Representative wage in 1976 (interview year)}
\item{hushrs}{Hours worked by husband in 1975}
\item{husage}{Husband's age in years}
\item{huseduc}{Husband's years of education}
\item{huswage}{Husband's estimated average hourly earnings}
\item{faminc}{Total family income}
\item{mtr}{Woman's Federal marginal income tax rate}
\item{motheduc}{Mother's years of education}
\item{fatheduc}{Father's years of education}
\item{unem}{Unemployment rate in county of residence}
\item{city}{Binary: lives in Standard Metropolitan Statistical Area (SMSA)}
\item{exper}{Years of labor market experience}
\item{nwifeinc}{Income from other sources than the wife's labor, in thousands}
\item{lwage}{\code{log(wage)}}
\item{expersq}{\code{exper^2}}
}
}
\source{
\url{https://cran.r-project.org/package=wooldridge}
}
\usage{
mroz
}
\description{
Sample of 753 between the ages of 30 and 60 in 1975 (PSID interview year 1976)
}
\references{
Wooldridge, J. M. (2020). \emph{Introductory Econometrics: A Modern Approach}.
7th Edition. Boston, MA: Cengage Learning.

Mroz, T. A. (1987). "The Sensitivity of an Empirical Model of Married Women's
Hours of Work to Economic and Statistical Assumptions."
\emph{Econometrica} 55, 765-799.
}
\keyword{datasets}
