% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlmodel-utils.R
\name{nobs.mlmodel}
\alias{nobs.mlmodel}
\title{Extract the Number of Observations from an mlmodel}
\usage{
\method{nobs}{mlmodel}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"mlmodel"}.}

\item{...}{Further arguments passed to methods (currently not used).}
}
\value{
An integer giving the number of observations used in the estimation
(after removing missing values and applying any \code{subset}).
}
\description{
Extract the Number of Observations from an mlmodel
}
