% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlmodel-count-tests.R
\name{OVDtest}
\alias{OVDtest}
\title{Overdispersion tests for count outcomes.}
\usage{
OVDtest(object)
}
\arguments{
\item{object}{An object of class \code{'mlmodel.count'}, i.e estimated by one of
our count variable models (ml_poisson, ml_negbin,...)}
}
\value{
A list with the results of overdispersion tests of a poisson model
against both a NB1 and NB2 negative binomial models.
}
\description{
Performs Cameron and Trivedi's test for overdispersion.
}
\references{
Cameron, A. C. and Trivedi, P. K. (1990). Regression-based tests for overdispersion in
the Poisson model. \emph{Journal of Econometrics}, 46(3), 347-364.

Cameron, A. C. and Trivedi, P. K. (2013). \emph{Regression Analysis of Count Data}.
2nd Edition. Cambridge University Press.
}
\seealso{
\code{\link{IMtest}} for general misspecification testing.
}
