% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlmodels-auxiliary.R
\name{parse_string_constraints}
\alias{parse_string_constraints}
\title{Internal helper: parse string constraints into maxLik matrices}
\usage{
parse_string_constraints(strings, coef_names)
}
\description{
Internal helper: parse string constraints into maxLik matrices
}
\keyword{internal}
