% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlmodel-tests.R
\name{print.IMtest.mlmodel}
\alias{print.IMtest.mlmodel}
\title{Prints the results of an object of class \code{IMtest.mlmodel}}
\usage{
\method{print}{IMtest.mlmodel}(x, digits = 4, ...)
}
\description{
Prints the results of an object of class \code{IMtest.mlmodel}
}
