% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm-postestimation.R
\name{print.summary.ml_lm}
\alias{print.summary.ml_lm}
\title{Print the summary statistics from an \code{ml_lm} estimation.}
\usage{
\method{print}{summary.ml_lm}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{An object of class \code{summary.ml_lm}, usually from \link[=summary.ml_lm]{summary.ml_lm()}.}

\item{digits}{A numeric scalar with the number of decimal places to use in
the statistics}

\item{...}{Currently not implemented.}
}
\description{
Print the summary statistics from an \code{ml_lm} estimation.
}
\author{
Alfonso Sanchez-Penalver
}
