% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlmodel-tests.R
\name{print.vuongtest.mlmodel}
\alias{print.vuongtest.mlmodel}
\title{Prints the result of an object of class \code{vuongtest.mlmodel}}
\usage{
\method{print}{vuongtest.mlmodel}(x, digits = 4, ...)
}
\description{
Prints the result of an object of class \code{vuongtest.mlmodel}
}
