% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlmodels-tests.R
\name{print.waldtest.mlmodel}
\alias{print.waldtest.mlmodel}
\title{Print method for Wald test results}
\usage{
\method{print}{waldtest.mlmodel}(x, digits = 3, ...)
}
\arguments{
\item{x}{An object of class \code{"waldtest.mlmodel"}, usually the output of
\code{\link[=waldtest]{waldtest()}}.}

\item{digits}{Integer. Number of significant digits to display for the
Wald statistic. Default is 3.}

\item{...}{Further arguments passed to or from other methods (currently
not used).}
}
\value{
The input object \code{x} (invisibly).
}
\description{
Prints a nicely formatted summary of the results from a Wald test
performed with \code{\link[=waldtest]{waldtest()}}.
}
\seealso{
\code{\link[=waldtest]{waldtest()}}
}
