% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pw401k}
\alias{pw401k}
\title{401(k) Participation Rates}
\format{
A data frame with 4734 rows and 5 variables:
\describe{
\item{prate}{participation rate, proportion}
\item{mrate}{matching rate, proportion}
\item{totemp}{total number of employees}
\item{age}{years the plan has been in place}
\item{sole}{binary indicating if it's sole plan offered by employer}
}
}
\source{
Papke, L. E. and Wooldridge, J. M. (1996).
}
\usage{
pw401k
}
\description{
A dataset containing participation rates and firm characteristics used
in Papke and Wooldridge (1996).
}
\references{
Papke, L. E., & Wooldridge, J. M. (1996). "Econometric methods for fractional
response variables with an application to 401(k) plan participation rates."
\emph{Journal of Applied Econometrics}, 11(6), 619-632.
\doi{10.1002/(SICI)1099-1255(199611)11:6<619::AID-JAE418>3.0.CO;2-1}
}
\keyword{datasets}
