% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm-postestimation.R
\name{residuals.ml_lm}
\alias{residuals.ml_lm}
\title{Get the residuals from a linear model estimated with \link{ml_lm}.}
\usage{
\method{residuals}{ml_lm}(object, ...)
}
\arguments{
\item{object}{A \code{ml_lm} object, from an estimation with \link{ml_lm}.}

\item{...}{Currently not used.}
}
\description{
Get the residuals from a linear model estimated with \link{ml_lm}.
}
\author{
Alfonso Sanchez-Penalver
}
