% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginaleffects-compat.R
\name{set_coef.mlmodel}
\alias{set_coef.mlmodel}
\title{Implementation of \code{marginaleffects} package's \code{set_coef()} function.}
\usage{
set_coef.mlmodel(model, coefs, ...)
}
\arguments{
\item{model}{An \code{mlmodel} estimation object (from one of the package's
estimation functions).}

\item{coefs}{A vector with the coefficients you want to se in the model.}

\item{...}{Additional arguments sent to methods.}
}
\value{
A new \code{mlmodel} object, with the coefficients set to those in
\code{coefs}/
}
\description{
Implementation of \code{marginaleffects} package's \code{set_coef()} function.
}
