% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smoke}
\alias{smoke}
\title{1979 National Health Interview Survey}
\format{
A data frame with 807 observations and 10 variables.
\describe{
\item{educ}{Years of education}
\item{cigpric}{State's cigarette price (cents per pack)}
\item{white}{Binary: white}
\item{age}{Age in years}
\item{income}{Annual income in dollars}
\item{cigs}{Number of cigarettes smoked per day}
\item{restaurn}{Binary: restaurant restrictions on smoking}
\item{lincome}{\code{log(income)}}
\item{agesq}{\code{age^2}}
\item{lcigpric}{\code{log(cigprice)}}
}
}
\source{
\url{https://cran.r-project.org/package=wooldridge}
}
\usage{
smoke
}
\description{
Sample of males in 1979 and early 1980 from the smoking supplement.
}
\references{
Wooldridge, J. M. (2020). \emph{Introductory Econometrics: A Modern Approach}.
7th Edition. Boston, MA: Cengage Learning.

Mullahy, J. (1997). "Instrumental-Variable Estimation of Count Data Models:
Applications to Models of Cigarette Smoking and Infant Health."
\emph{Review of Economics and Statistics} 79, 586-593.
}
\keyword{datasets}
