% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson-postestimation.R
\name{summary.ml_poisson}
\alias{summary.ml_poisson}
\title{Summary for ml_poisson objects}
\usage{
\method{summary}{ml_poisson}(
  object,
  correlation = FALSE,
  vcov = NULL,
  vcov.type = "oim",
  cl_var = NULL,
  repetitions = 999,
  seed = NULL,
  progress = FALSE,
  ...
)
}
\arguments{
\item{object}{A fitted model object of class \code{"ml_poisson"}.}

\item{correlation}{Logical. Should the correlation matrix of the estimated
parameters be included in the output? Default is \code{FALSE}. If \code{TRUE} the
correlation matrix will be computed, and stored in the \code{'summary.ml_lm'}
object the function returns.}

\item{vcov}{Optional user-supplied variance-covariance matrix. If provided,
it will be used instead of computing one internally.}

\item{vcov.type}{Character string specifying the type of variance-covariance
matrix to use. See \link[=vcov.mlmodel]{vcov}.}

\item{cl_var}{Character string or vector. Name of the clustering variable
or the vector itself. See \link[=vcov.mlmodel]{vcov}.}

\item{repetitions}{Integer. Number of bootstrap replications when
\code{vcov.type = "boot"}. Default is 999.}

\item{seed}{Integer. Random seed for reproducibility when \code{vcov.type = "boot"}.
If \code{NULL}, a random seed is generated.}

\item{progress}{Logical. Should a progress bar be displayed during
bootstrapping or jackknifing? Default is \code{FALSE} (silent).}

\item{...}{Further arguments passed to methods.}
}
\value{
An object of class \code{c("summary.ml_poisson", "summary.mlmodel", "summary")}.
}
\description{
Summary for ml_poisson objects
}
\details{
Coefficient names in the fitted object use the prefixes \verb{value::} for
consistency with our other estimators that model the scale as well.
}
\author{
Alfonso Sanchez-Penalver
}
