% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlmodel-utils.R
\name{terms.mlmodel}
\alias{terms.mlmodel}
\title{Extract terms from mlmodel objects}
\usage{
\method{terms}{mlmodel}(x, ...)
}
\arguments{
\item{x}{An \code{mlmodel} object from an estimation of one of our models.}

\item{...}{Not currently implemented.}
}
\description{
Returns the terms object for the value (mean/probability) equation.
For heteroskedastic models, the scale equation is \strong{not} included in the
returned terms object.
}
\note{
The \code{terms()} method is provided for minimal compatibility with other
packages, but it is incomplete for heteroskedastic models. For reliable
bootstrap standard errors, use \code{vcov(object, type = "boot")} instead of
functions from the sandwich package.
}
