% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson-postestimation.R
\name{update.ml_poisson}
\alias{update.ml_poisson}
\title{Update method for ml_poisson objects}
\usage{
\method{update}{ml_poisson}(
  object,
  formula. = NULL,
  data = NULL,
  weights = NULL,
  ...,
  evaluate = TRUE
)
}
\arguments{
\item{object}{An \code{ml_poisson} estimation object.}

\item{formula.}{The formula of the value equation (optional).}

\item{data}{A data.frame with the data to do the estimation (optional).}

\item{weights}{A vector with the weights (optional).}

\item{...}{Currently not implemented.}

\item{evaluate}{Should the updated call be evaluated? Defaults to \code{TRUE}.}

\item{scale.}{The formula of the scale equation (optional).}
}
\description{
Update method for ml_poisson objects
}
\details{
Re-evaluates the original call with new data/weights while preserving
the original scale formula, constraints, control, etc.

\strong{Note on weights}: If the original model was weighted, sandwich
usually passes weights = NULL. In that case we keep the original weights.
This means the bootstrap may not be properly re-weighted. For accurate
weighted bootstrap use our own \link[=vcov.mlmodel]{vcov()} instead.

\strong{Note on sandwich}: This package's functions does not work reliably with
\code{ml_poisson} objects.  We, therefore, built our own bootstrap and jackknife
implementations. We strongly recommend using \link[=vcov.mlmodel]{vcov()} instead.
}
