% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlmodels-auxiliary.R
\name{vcov_cluster_info}
\alias{vcov_cluster_info}
\title{Internal helper to extract clustering information}
\usage{
vcov_cluster_info(object, cl_var)
}
\arguments{
\item{object}{An \code{mlmodel} object.}

\item{cl_var}{Character string or vector. The clustering variable.}
}
\value{
A list with \code{var_name}, \code{n_cluster}, and \code{ids}.
}
\description{
Used by \code{vcov.mlmodel()} when \code{vcov.type = "cluster"} or \code{"cluster-boot"}.
}
\keyword{internal}
