% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlmodel-utils.R
\name{BIC.mlmodel}
\alias{BIC.mlmodel}
\title{Extract BIC}
\usage{
\method{BIC}{mlmodel}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"mlmodel"} or \code{"summary.mlmodel"}.}

\item{...}{Further arguments passed to methods.}
}
\description{
Extract BIC
}
