% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlmodel-utils.R
\name{BIC.mlmodel}
\alias{BIC.mlmodel}
\alias{BIC.summary.mlmodel}
\title{Extract BIC from mlmodel objects}
\usage{
\method{BIC}{mlmodel}(object, ...)

\method{BIC}{summary.mlmodel}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"mlmodel"} or \code{"summary.mlmodel"}.}

\item{...}{Further arguments passed to methods.}
}
\value{
A numeric value with the BIC.
}
\description{
Extract BIC from mlmodel objects
}
\details{
BIC is computed as \code{-2 * logLik(object) + log(nobs) * npar}.
}
