% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlmodel-utils.R
\name{coef.mlmodel}
\alias{coef.mlmodel}
\title{Gets the coefficients from an mlmodel estimation.}
\usage{
\method{coef}{mlmodel}(object, ...)
}
\arguments{
\item{object}{An mlmodel from an estimator, that you want the coefficients for.}

\item{...}{Currently not used.}
}
\value{
A vector with the coefficients from the estimation.
}
\description{
Gets the coefficients from an mlmodel estimation.
}
\author{
Alfonso Sanchez-Penalver
}
