% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlmodels-auxiliary.R
\name{detect_log_transformation}
\alias{detect_log_transformation}
\title{Detect log transformation on the outcome variable}
\usage{
detect_log_transformation(formula, data)
}
\arguments{
\item{formula}{a formula with a lhs expression to check for.}

\item{data}{A data.frame that holds the variables in the estimation.}
}
\value{
A list with all the information of a log transformation.
}
\description{
Returns a consistent list structure whether a log transformation is detected or not.
}
\keyword{internal}
