% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlmodels-auxiliary.R
\name{detect_log_transformations}
\alias{detect_log_transformations}
\title{Detect log transformations across multiple formulas}
\usage{
detect_log_transformations(formulas_list, data)
}
\arguments{
\item{data}{A data.frame that holds all the data in the estimation.}

\item{formulas}{A list with the different formulas for the different
equations that need checking.}
}
\value{
A list with the same number of elements as in \code{formulas_list}, each
with an inner list with all the information about the log-transformation for
the respective equation.
}
\description{
Returns a named list with detection results for each formula.
Useful for future multi-equation models.
}
\keyword{internal}
