% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlmodels-auxiliary.R
\name{.mold_fix_integer_to_double}
\alias{.mold_fix_integer_to_double}
\title{Function to change ptypes in a mold from int to double.}
\usage{
.mold_fix_integer_to_double(mold)
}
\arguments{
\item{mold}{An object returned by \link[hardhat:mold]{mold()}.}
}
\value{
The \code{mold} object with the changed \code{ptype} tibble.
}
\description{
Function to change ptypes in a mold from int to double.
}
\keyword{internal}
