% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma-auxiliary.R, R/mlmodel-auxiliary.R
\name{.vcov_boot.ml_gamma}
\alias{.vcov_boot.ml_gamma}
\alias{.vcov_boot}
\alias{.vcov_boot.mlmodel}
\title{Bootstrap Variance-Covariance Matrix (mlmodel method)}
\usage{
\method{.vcov_boot}{ml_gamma}(
  object,
  repetitions = 999,
  seed = NULL,
  cl_var = NULL,
  progress = TRUE,
  ...
)

.vcov_boot(object, ...)

\method{.vcov_boot}{mlmodel}(
  object,
  repetitions = 999,
  seed = NULL,
  cl_var = NULL,
  progress = TRUE,
  ...
)
}
\arguments{
\item{object}{An \code{mlmodel} object.}

\item{repetitions}{Number of bootstrap replications.}

\item{seed}{Random seed for reproducibility.}

\item{cl_var}{Clustering variable (if clustered bootstrap).}

\item{progress}{Logical. Whether to show progress bar.}

\item{...}{Not currently used.}
}
\description{
Internal function to compute bootstrapped variance-covariance matrix.
}
\details{
Called by \link[=vcov]{vcov} when \code{type} is set to \code{"boot"}
}
\keyword{internal}
