% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlmodel-utils.R
\name{fitted.mlmodel}
\alias{fitted.mlmodel}
\alias{fitted.values.mlmodel}
\title{Returns the fitted values from an object of class \code{mlmodel}.}
\usage{
\method{fitted}{mlmodel}(object, ...)

\method{fitted}{values.mlmodel}(object, ...)

\method{fitted}{mlmodel}(object, ...)
}
\arguments{
\item{object}{An \code{mlmodel} object.}

\item{...}{Further arguments passed to methods (currently ignored).}
}
\value{
A vector with the fitted values of the model.

A numeric vector of fitted values, aligned to the original sample size.
Observations that were dropped (due to \code{subset}, \code{NA}s, etc.) will be \code{NA}.
}
\description{
Returns the fitted values from an object of class \code{mlmodel}.

Extract Fitted Values
}
