% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginaleffects-compat.R
\name{get_data.mlmodel}
\alias{get_data.mlmodel}
\alias{get_modeldata.mlmodel}
\title{Implementation of \code{insight} package's \code{get_data()} function.}
\usage{
\method{get_data}{mlmodel}(x, ...)

get_modeldata.mlmodel(x, ...)
}
\arguments{
\item{x}{An \code{mlmodel} estimation object (from one of the package's
estimation functions).}

\item{...}{Additional arguments sent to methods.}
}
\value{
A data.frame with the data the user passed when estimating the
model.
}
\description{
Implementation of \code{insight} package's \code{get_data()} function.
}
