% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlmodel-utils.R
\name{gradientObs}
\alias{gradientObs}
\alias{gradientObs.mlmodel}
\title{Gradient by observation}
\usage{
gradientObs(object)

\method{gradientObs}{mlmodel}(object)
}
\arguments{
\item{object}{An object of class \code{mlmodel}, i.e. an model estimated with one
of the estimators in the package.}
}
\value{
Matrix with each observation's gradient (score) in a row.
}
\description{
Get the gradients (scores), evaluated at the optimum from an object of class
\code{mlmodel}.
}
\author{
Alfonso Sanchez-Penalver
}
