% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlmodel-utils.R
\name{hessianObs}
\alias{hessianObs}
\alias{hessianObs.mlmodel}
\title{Hessian matrix by observation}
\usage{
hessianObs(object)

\method{hessianObs}{mlmodel}(object)
}
\arguments{
\item{object}{An object of class \code{mlmodel}, i.e. an model estimated with one
of the estimators in the package.}
}
\value{
Matrix with the evaluated Hessian matrices per observation.
}
\description{
Get the hessian matrix, evaluated at the optimum, by observation from an
object of class \code{mlmodel}.
}
\details{
If there are N observations in the estimation, and the model has K parameters,
this method returns a (N * K) by K matrix. The Hessian for each observation
stacked on top of each other.
}
\author{
Alfonso Sanchez-Penalver
}
