% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit-fit.R
\name{ml_logit}
\alias{ml_logit}
\title{Fit Binary Logit Model by Maximum Likelihood}
\usage{
ml_logit(
  value,
  scale = NULL,
  weights = NULL,
  data,
  subset = NULL,
  noint_value = FALSE,
  control = list(tol = -1, reltol = 1e-12, gradtol = 1e-12, lambdatol = 1e-20, qac =
    "marquardt"),
  ...
)
}
\arguments{
\item{value}{Two-sided formula for the probability equation.}

\item{scale}{Optional one-sided formula for heteroskedasticity.}

\item{weights}{Optional weights.}

\item{data}{Data frame.}

\item{subset}{Optional subset expression. Only observations for which this
expression evaluates to \code{TRUE} are used in the estimation. This can be
a logical vector or an expression (e.g. \code{subset = age > 30}).}

\item{noint_value}{Logical. Should the value equation omit the intercept?
Default is \code{FALSE}.}

\item{control}{A list of control parameters passed to \code{\link[maxLik:maxLik]{maxLik::maxLik()}}.
The default values are chosen to work well with difficult likelihoods.
See \code{\link[maxLik:maxLik]{maxLik::maxLik()}} for details.}

\item{...}{Additional arguments passed to \code{\link[maxLik:maxLik]{maxLik::maxLik()}}.}
}
\value{
An object of class \code{ml_logit} that extends \code{mlmodel}.
}
\description{
Fit Binary Logit Model by Maximum Likelihood
}
\details{
\strong{Important:} Do not use the usual R syntax to remove the intercept in the
formula (\code{- 1} or \code{+ 0}) for the value equation. Use the dedicated argument
\code{noint_value} instead. Remember that for the scale, we need a formula with
only the right hand side part.
}
\author{
Alfonso Sanchez-Penalver
}
