% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlmodels-package.R
\docType{package}
\name{mlmodels-package}
\alias{mlmodels}
\alias{mlmodels-package}
\title{mlmodels: Maximum Likelihood Models and Tools for Estimation, Prediction, and Testing}
\description{
A collection of tools for estimating statistical models via maximum likelihood estimation (MLE). The package provides flexible model specifications, methods for generating predictions, and tools for hypothesis testing and inference. It builds on established optimization routines while offering a consistent, user-friendly interface for custom and standard maximum likelihood models.
}
\author{
\strong{Maintainer}: Alfonso Sanchez-Penalver \email{oneiros_spain@yahoo.com} (\href{https://orcid.org/0000-0001-8491-4632}{ORCID})

}
\keyword{internal}
