% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlmodel-utils.R
\name{nobs.mlmodel}
\alias{nobs.mlmodel}
\title{Returns the number of observations used in an estimation of an \code{mlmodel} model.}
\usage{
\method{nobs}{mlmodel}(object, ...)
}
\arguments{
\item{object}{An \code{mlmodel} estimation model.}

\item{...}{Not currently implemented.}
}
\value{
A value with the number of observations used in the estimation.
}
\description{
Returns the number of observations used in an estimation of an \code{mlmodel} model.
}
