% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlmodels-utils.R
\name{predict.mlmodel}
\alias{predict.mlmodel}
\title{Predictions for mlmodel objects}
\usage{
\method{predict}{mlmodel}(object, ...)
}
\arguments{
\item{object}{A model object inheriting from \code{'mlmodel'}.}

\item{...}{Further arguments passed to specific methods.}
}
\description{
Generic method for computing predictions from models fitted with the
\code{mlmodels} package.
}
