% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlmodel-tests.R
\name{print.vuongtest.mlmodel}
\alias{print.vuongtest.mlmodel}
\title{Prints the result of an object of class \code{vuongtest.mlmodel}}
\usage{
\method{print}{vuongtest.mlmodel}(x, digits = 4, ...)
}
\arguments{
\item{x}{An object of class \code{vuongtest.mlmodel}, i.e. from \code{vuongtest()}}

\item{digits}{Number of decimal places to display (not used)}

\item{...}{Currently not implemented}
}
\description{
Prints the result of an object of class \code{vuongtest.mlmodel}
}
