% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm-postestimation.R
\name{residuals.ml_lm}
\alias{residuals.ml_lm}
\title{Extract Residuals from ml_lm objects}
\usage{
\method{residuals}{ml_lm}(object, ...)
}
\arguments{
\item{object}{A fitted \code{ml_lm} object.}

\item{...}{Not currently used.}
}
\value{
A numeric vector of residuals with length equal to the number of
observations used in estimation.
}
\description{
Returns the residuals (observed minus fitted values) from the value equation.
}
\author{
Alfonso Sanchez-Penalver
}
