% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm-postestimation.R, R/logit-postestimation.R,
%   R/mlmodel-utils.R, R/negbin-postestimation.R, R/poisson-postestimation.R,
%   R/probit-postestimation.R
\name{update.ml_lm}
\alias{update.ml_lm}
\alias{update.ml_logit}
\alias{update.mlmodel}
\alias{update.ml_negbin}
\alias{update.ml_poisson}
\alias{update.ml_probit}
\title{Update method for ml_lm objects}
\usage{
\method{update}{ml_lm}(
  object,
  formula. = NULL,
  scale. = NULL,
  data = NULL,
  weights = NULL,
  ...,
  evaluate = TRUE
)

\method{update}{ml_logit}(
  object,
  formula. = NULL,
  scale. = NULL,
  data = NULL,
  weights = NULL,
  ...,
  evaluate = TRUE
)

update.mlmodel(
  object,
  formula. = NULL,
  scale. = NULL,
  data = NULL,
  weights = NULL,
  ...,
  evaluate = TRUE
)

\method{update}{ml_negbin}(
  object,
  formula. = NULL,
  data = NULL,
  weights = NULL,
  ...,
  evaluate = TRUE
)

\method{update}{ml_poisson}(
  object,
  formula. = NULL,
  data = NULL,
  weights = NULL,
  ...,
  evaluate = TRUE
)

\method{update}{ml_probit}(
  object,
  formula. = NULL,
  scale. = NULL,
  data = NULL,
  weights = NULL,
  ...,
  evaluate = TRUE
)
}
\arguments{
\item{object}{An \code{mlmodel} estimation object.}

\item{formula.}{The formula of the value equation (optional).}

\item{scale.}{The formula of the scale equation (optional).}

\item{data}{A data.frame with the data to do the estimation (optional).}

\item{weights}{A vector with the weights (optional).}

\item{...}{Currently not implemented.}

\item{evaluate}{Should the updated call be evaluated? Defaults to \code{TRUE}.}
}
\description{
Update method for ml_lm objects
}
\details{
Used in bootstrapping of tests. Re-evaluates the original call with arguments
passed into the function.

\strong{Note on sandwich::vcovBS()}: This function does not work reliably with
\code{mlmodel} objects, even in simple homoskedastic cases.  We, therefore, built
our own bootstrap implementation. We strongly recommend using
\code{vcov(object, type = "boot")} instead. See \link[=vcov.mlmodel]{vcov}.
}
\author{
Alfonso Sanchez-Penalver
}
