% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlmodel-tests.R
\name{vuongtest}
\alias{vuongtest}
\alias{vuongtest.mlmodel}
\title{Vuong's Test for \code{mlmodel} objects}
\usage{
vuongtest(object_1, object_2, ...)

\method{vuongtest}{mlmodel}(object_1, object_2, ...)
}
\arguments{
\item{object_1}{A fitted model object inheriting from \code{"mlmodel"}.}

\item{object_2}{A fitted model object inheriting from \code{"mlmodel"}.}

\item{...}{Further arguments passed to methods (not currently implemented).}
}
\value{
An object of class \code{vuongtest.mlmodel}
}
\description{
Performs Vuong's test between two models estimated via maximum likelihood.
}
\references{
Vuong, Q. H. (1989). Likelihood Ratio Tests for Model Selection and Non-Nested
Hypotheses. \emph{Econometrica}, 57(2), 307-333. \doi{10.2307/1912557}
}
\author{
Alfonso Sanchez-Penalver
}
